[update]拐角过渡完善
圆形，延长相交的实现

+++ libtess2/Include/tesselator.h
@@ -129,8 +129,7 @@ enum TessOption
 	TESS_REVERSE_CONTOURS
 };
 
-//typedef float TESSreal;
-typedef double TESSreal;
+typedef float TESSreal;
 typedef int TESSindex;
 typedef struct TESStesselator TESStesselator;
 typedef struct TESSalloc TESSalloc;
